+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Detect running AWS and Azure clusters using `Cluster` CRs instead of provider-specific CRs
+
 ## [2.1.0] - 2020-09-16
 
 ### Added